fix(api-page-builder-import-export): add support for file keys with sub-folders
