+++ packages/api-page-builder-import-export/src/import/constants.ts
@@ -1,0 +1,5 @@
+import path from "path";
+
+export const INSTALL_DIR = "/tmp";
+export const INSTALL_EXTRACT_DIR = path.join(INSTALL_DIR, "apiPageBuilderImport");
+export const ASSETS_DIR_NAME = "/assets";
