+++ packages/api-page-builder-import-export/src/import/create/formsHandler.ts
@@ -1,10 +1,11 @@
 import { ImportExportTaskStatus, PbImportExportContext } from "~/types";
-import { initialStats, readExtractAndUploadZipFileContents } from "~/import/utils";
+import { initialStats } from "~/import/utils";
 import { invokeHandlerClient } from "~/client";
 import { Payload as ProcessPayload } from "../process";
 import { mockSecurity } from "~/mockSecurity";
 import { zeroPad } from "@webiny/utils";
 import { Configuration, Payload, Response } from "~/import/create";
+import { extractAndUploadZipFileContents } from "~/import/utils/extractAndUploadZipFileContents";
 
 export const formsHandler = async (
     configuration: Configuration,
@@ -27,12 +28,12 @@ export const formsHandler = async (
         }
         mockSecurity(identity, context);
         // Step 1: Read the zip file
-        const formImportDataList = await readExtractAndUploadZipFileContents(zipFileUrl);
+        const formImportDataList = await extractAndUploadZipFileContents(zipFileUrl);
 
         // For each form create a subtask and invoke the process handler
         for (let i = 0; i < formImportDataList.length; i++) {
             const formsDirMap = formImportDataList[i];
-            // Create sub task
+            // Create sub-task
             const subtask = await pageBuilder.importExportTask.createSubTask(
                 task.id,
                 zeroPad(i + 1, 5),
