+++ packages/api-page-builder-import-export/src/import/utils/getFileNameWithoutExt.ts
@@ -1,0 +1,5 @@
+import path from "path";
+
+export function getFileNameWithoutExt(fileName: string): string {
+    return path.basename(fileName).replace(path.extname(fileName), "");
+}
